🎉 feat: Add safe_filename helper function
The `safe_filename` function has been added to the `src/helpers/safe_filename.rs` module. This function takes a string `s` and an optional `max_length` parameter and returns a sanitized version of the input string that can be used as a safe filename.

The function removes characters that are not allowed in NTFS filenames (characters in the range 0-31) as well as other invalid characters such as `"`, `#`, `$`, `%`, etc. The function also truncates the resulting filename to the specified `max_length` if provided.

The `safe_filename` function is useful when generating filenames from user-provided or dynamic content to ensure they are safe and compatible with file systems.

Additionally, the `rstest` crate has been added to the `dev-dependencies` section in the `Cargo.toml` file, which indicates that it is used for testing purposes.

Signed-off-by: Sabry Awad <dr.sabry1997@gmail.com>